Remove .newPublishingSheet FF
