+++ WordPress/Classes/Utility/BuildInformation/FeatureFlag.swift
@@ -26,7 +26,6 @@ public enum FeatureFlag: Int, CaseIterable {
     case pluginManagementOverhaul
     case newsletterSubscribers
     case newStats
-    case newPublishingSheet
     case mediaQuotaView
     case intelligence
 
@@ -85,8 +84,6 @@ public enum FeatureFlag: Int, CaseIterable {
             return true
         case .newStats:
             return false
-        case .newPublishingSheet:
-            return true
         case .mediaQuotaView:
             return false
         case .intelligence:
@@ -135,7 +132,6 @@ extension FeatureFlag {
         case .readerGutenbergCommentComposer: "Gutenberg Comment Composer"
         case .newsletterSubscribers: "Newsletter Subscribers"
         case .newStats: "New Stats"
-        case .newPublishingSheet: "New Publishing Sheet"
         case .mediaQuotaView: "Media Quota"
         case .intelligence: "Intelligence"
         }
